dlm_sampling

This is function draws samples from the latent states using the backward sampling algorithm. See West and Harrison (1997) , chapter 15, for details.

Usage

dlm_sampling(model, sample.size, filt.distr = FALSE)

Arguments

model

fitted_dlm: A fitted model from which to sample.

sample.size

integer: The number of samples to draw.

filt.distr

boolean: A flag indicating if the samples should be dawned from the filtered (if TRUE) or smoothed distribution. The default is FALSE.

Value

A list containing the following values:

mt array: An array containing a sample of the latent states. Dimensions are n x t x sample.size, where n is the number of latent states in the model and t is the number of observed values.

ft array: An array containing a sample of the linear predictors. Dimensions are k x t x sample.size, where k is the number of linear predictors in the model and t is the number of observed values.

param list: A named list containing, for each model outcome, an array with the samples of the parameters of the observational model. Each array will have dimensions l x t x sample.size, where l is the number of parameters in the observational model and t is the number of observed values.

See also

Examples

structure <- polynomial_block(mu = 1, D = 0.95) +
  polynomial_block(V = 1, D = 0.95)

outcome <- Normal(mu = "mu", V = "V", data = cornWheat$corn.log.return[1:500])
fitted.data <- fit_model(structure, corn = outcome)

sample <- dlm_sampling(fitted.data, 5000)