Search the best hyper parameters fo a kDGLM model

Search the best hyper parameters fo a kDGLM model

Usage

search_model(
  ...,
  search.grid,
  condition = "TRUE",
  metric = "log.like",
  smooth = TRUE,
  lag = 1,
  pred.cred = 0.95,
  metric.cutoff = NA,
  p.monit = NA,
  save.models = FALSE,
  silent = FALSE
)

Arguments

...

dlm_block or dlm_distr objects: The structural blocks of the model (dlm_block objects), alongside the model outcomes (dlm_distr object). At least one block must be undefined.

search.grid

list: A named list containing the possible values of each undefined hyper parameter.

condition

character: A character defining which combinations of undefined hyper parameter should be tested. See example for details.

metric

character: The name of the metric to use for model selection. One of log-likelihood for the one-step-ahead prediction ("log.like"), Mean Absolute Error ("mae"), Mean Absolute Scaled Error ("mase") (Hyndman and Koehler 2006) , Relative Absolute Error ("rae"), Mean Squared Error ("mse") or Interval Score ("interval.score") (Bracher et al. 2022) .

smooth

boolean: A flag indicating if the smoothed distribution for the latent states should be calculated.

lag

integer: The number of steps ahead used for the prediction when calculating the metrics. If lag<0, predictions are made using the smoothed distribution of the latent states.

pred.cred

numeric: A number between 0 and 1 (not included) indicating the credibility interval for predictions. If not within the valid range of values, 0.95 will be used.

metric.cutoff

integer: The number of observations to ignore when calculating the metrics. Default is 1/10 of the number of observations (rounded down).

p.monit

numeric (optional): The prior probability of changes in the latent space variables that are not part of its dynamic.

save.models

boolean: A flag indicating if all evaluated models should be saved. If FALSE, only the best model (according to the chosen metric) will be saved;

silent

boolean: A flag indicating if a progress bar should be printed.

Value

A searched_dlm object containing the following values:

search.data data.frame: A table contain the metrics of each combination of hyper parameter tested and ordered from best combination to worst.

structure dlm_block object: The initial structure passed by the user, by with the undefined hyper parameters set to the best values found.

model fitted_dlm object: A model fitted with the best combination of hyper parameters.

Details

This is an auxiliary function to search for the best combination of hyper parameters among the possible variations specified by the user. This function simple evaluates all possible combinations and chooses the one that results in the best fitting.

References

Johannes Bracher, Evan L. Ray, Tilmann Gneiting, Nicholas G. Reich (2022). “Correction: Evaluating epidemic forecasts in an interval format.” PLOS Computational Biology, 18(10), 1-3. doi:10.1371/journal.pcbi.1010592 .

Rob J. Hyndman, Anne B. Koehler (2006). “Another look at measures of forecast accuracy.” International Journal of Forecasting, 22(4), 679--688. http://www.sciencedirect.com/science/article/B6V92-4K18VK5-1/1/227cef013b86100791c53712fae7677c.

See also

Examples


data <- c(AirPassengers)

level <- polynomial_block(rate = 1, order = 2, D = "D.level")
season <- harmonic_block(rate = "sazo.effect", order = 2, period = 12, D = "D.sazo")

outcome <- Poisson(lambda = "rate", data = data)

search_model(level, season,
  outcomes = outcome,
  search.grid = list(
    sazo.effect = c(0, 1),
    D.level = c(seq.int(0.8, 1, l = 11)),
    D.sazo = c(seq.int(0.95, 1, l = 11))
  ),
  condition = "sazo.effect==1 | D.sazo==1"
)
#> [                                                  ] - 0.76% - ETA - 0 minutes                 [=                                                 ] - 1.52% - ETA - 0.08 minutes                 [=                                                 ] - 2.27% - ETA - 0.1 minutes                 [==                                                ] - 3.03% - ETA - 0.11 minutes                 [==                                                ] - 3.79% - ETA - 0.12 minutes                 [==                                                ] - 4.55% - ETA - 0.13 minutes                 [===                                               ] - 5.3% - ETA - 0.13 minutes                 [===                                               ] - 6.06% - ETA - 0.12 minutes                 [===                                               ] - 6.82% - ETA - 0.12 minutes                 [====                                              ] - 7.58% - ETA - 0.12 minutes                 [====                                              ] - 8.33% - ETA - 0.12 minutes                 [=====                                             ] - 9.09% - ETA - 0.12 minutes                 [=====                                             ] - 9.85% - ETA - 0.12 minutes                 [=====                                             ] - 10.61% - ETA - 0.12 minutes                 [======                                            ] - 11.36% - ETA - 0.12 minutes                 [======                                            ] - 12.12% - ETA - 0.12 minutes                 [======                                            ] - 12.88% - ETA - 0.11 minutes                 [=======                                           ] - 13.64% - ETA - 0.11 minutes                 [=======                                           ] - 14.39% - ETA - 0.12 minutes                 [========                                          ] - 15.15% - ETA - 0.11 minutes                 [========                                          ] - 15.91% - ETA - 0.11 minutes                 [========                                          ] - 16.67% - ETA - 0.11 minutes                 [=========                                         ] - 17.42% - ETA - 0.11 minutes                 [=========                                         ] - 18.18% - ETA - 0.11 minutes                 [=========                                         ] - 18.94% - ETA - 0.11 minutes                 [==========                                        ] - 19.7% - ETA - 0.11 minutes                 [==========                                        ] - 20.45% - ETA - 0.11 minutes                 [===========                                       ] - 21.21% - ETA - 0.11 minutes                 [===========                                       ] - 21.97% - ETA - 0.11 minutes                 [===========                                       ] - 22.73% - ETA - 0.11 minutes                 [============                                      ] - 23.48% - ETA - 0.11 minutes                 [============                                      ] - 24.24% - ETA - 0.11 minutes                 [============                                      ] - 25% - ETA - 0.1 minutes                 [=============                                     ] - 25.76% - ETA - 0.1 minutes                 [=============                                     ] - 26.52% - ETA - 0.1 minutes                 [==============                                    ] - 27.27% - ETA - 0.1 minutes                 [==============                                    ] - 28.03% - ETA - 0.1 minutes                 [==============                                    ] - 28.79% - ETA - 0.1 minutes                 [===============                                   ] - 29.55% - ETA - 0.1 minutes                 [===============                                   ] - 30.3% - ETA - 0.1 minutes                 [================                                  ] - 31.06% - ETA - 0.09 minutes                 [================                                  ] - 31.82% - ETA - 0.09 minutes                 [================                                  ] - 32.58% - ETA - 0.09 minutes                 [=================                                 ] - 33.33% - ETA - 0.09 minutes                 [=================                                 ] - 34.09% - ETA - 0.09 minutes                 [=================                                 ] - 34.85% - ETA - 0.09 minutes                 [==================                                ] - 35.61% - ETA - 0.09 minutes                 [==================                                ] - 36.36% - ETA - 0.09 minutes                 [===================                               ] - 37.12% - ETA - 0.09 minutes                 [===================                               ] - 37.88% - ETA - 0.09 minutes                 [===================                               ] - 38.64% - ETA - 0.08 minutes                 [====================                              ] - 39.39% - ETA - 0.08 minutes                 [====================                              ] - 40.15% - ETA - 0.08 minutes                 [====================                              ] - 40.91% - ETA - 0.08 minutes                 [=====================                             ] - 41.67% - ETA - 0.08 minutes                 [=====================                             ] - 42.42% - ETA - 0.08 minutes                 [======================                            ] - 43.18% - ETA - 0.08 minutes                 [======================                            ] - 43.94% - ETA - 0.08 minutes                 [======================                            ] - 44.7% - ETA - 0.08 minutes                 [=======================                           ] - 45.45% - ETA - 0.07 minutes                 [=======================                           ] - 46.21% - ETA - 0.07 minutes                 [=======================                           ] - 46.97% - ETA - 0.07 minutes                 [========================                          ] - 47.73% - ETA - 0.07 minutes                 [========================                          ] - 48.48% - ETA - 0.07 minutes                 [=========================                         ] - 49.24% - ETA - 0.07 minutes                 [=========================                         ] - 50% - ETA - 0.07 minutes                 [=========================                         ] - 50.76% - ETA - 0.07 minutes                 [==========================                        ] - 51.52% - ETA - 0.07 minutes                 [==========================                        ] - 52.27% - ETA - 0.07 minutes                 [===========================                       ] - 53.03% - ETA - 0.06 minutes                 [===========================                       ] - 53.79% - ETA - 0.06 minutes                 [===========================                       ] - 54.55% - ETA - 0.06 minutes                 [============================                      ] - 55.3% - ETA - 0.06 minutes                 [============================                      ] - 56.06% - ETA - 0.06 minutes                 [============================                      ] - 56.82% - ETA - 0.06 minutes                 [=============================                     ] - 57.58% - ETA - 0.06 minutes                 [=============================                     ] - 58.33% - ETA - 0.06 minutes                 [==============================                    ] - 59.09% - ETA - 0.06 minutes                 [==============================                    ] - 59.85% - ETA - 0.06 minutes                 [==============================                    ] - 60.61% - ETA - 0.05 minutes                 [===============================                   ] - 61.36% - ETA - 0.05 minutes                 [===============================                   ] - 62.12% - ETA - 0.05 minutes                 [===============================                   ] - 62.88% - ETA - 0.05 minutes                 [================================                  ] - 63.64% - ETA - 0.05 minutes                 [================================                  ] - 64.39% - ETA - 0.05 minutes                 [=================================                 ] - 65.15% - ETA - 0.05 minutes                 [=================================                 ] - 65.91% - ETA - 0.05 minutes                 [=================================                 ] - 66.67% - ETA - 0.05 minutes                 [==================================                ] - 67.42% - ETA - 0.04 minutes                 [==================================                ] - 68.18% - ETA - 0.04 minutes                 [==================================                ] - 68.94% - ETA - 0.04 minutes                 [===================================               ] - 69.7% - ETA - 0.04 minutes                 [===================================               ] - 70.45% - ETA - 0.04 minutes                 [====================================              ] - 71.21% - ETA - 0.04 minutes                 [====================================              ] - 71.97% - ETA - 0.04 minutes                 [====================================              ] - 72.73% - ETA - 0.04 minutes                 [=====================================             ] - 73.48% - ETA - 0.04 minutes                 [=====================================             ] - 74.24% - ETA - 0.04 minutes                 [======================================            ] - 75% - ETA - 0.03 minutes                 [======================================            ] - 75.76% - ETA - 0.03 minutes                 [======================================            ] - 76.52% - ETA - 0.03 minutes                 [=======================================           ] - 77.27% - ETA - 0.03 minutes                 [=======================================           ] - 78.03% - ETA - 0.03 minutes                 [=======================================           ] - 78.79% - ETA - 0.03 minutes                 [========================================          ] - 79.55% - ETA - 0.03 minutes                 [========================================          ] - 80.3% - ETA - 0.03 minutes                 [=========================================         ] - 81.06% - ETA - 0.03 minutes                 [=========================================         ] - 81.82% - ETA - 0.02 minutes                 [=========================================         ] - 82.58% - ETA - 0.02 minutes                 [==========================================        ] - 83.33% - ETA - 0.02 minutes                 [==========================================        ] - 84.09% - ETA - 0.02 minutes                 [==========================================        ] - 84.85% - ETA - 0.02 minutes                 [===========================================       ] - 85.61% - ETA - 0.02 minutes                 [===========================================       ] - 86.36% - ETA - 0.02 minutes                 [============================================      ] - 87.12% - ETA - 0.02 minutes                 [============================================      ] - 87.88% - ETA - 0.02 minutes                 [============================================      ] - 88.64% - ETA - 0.02 minutes                 [=============================================     ] - 89.39% - ETA - 0.01 minutes                 [=============================================     ] - 90.15% - ETA - 0.01 minutes                 [=============================================     ] - 90.91% - ETA - 0.01 minutes                 [==============================================    ] - 91.67% - ETA - 0.01 minutes                 [==============================================    ] - 92.42% - ETA - 0.01 minutes                 [===============================================   ] - 93.18% - ETA - 0.01 minutes                 [===============================================   ] - 93.94% - ETA - 0.01 minutes                 [===============================================   ] - 94.7% - ETA - 0.01 minutes                 [================================================  ] - 95.45% - ETA - 0.01 minutes                 [================================================  ] - 96.21% - ETA - 0.01 minutes                 [================================================  ] - 96.97% - ETA - 0 minutes                 [================================================= ] - 97.73% - ETA - 0 minutes                 [================================================= ] - 98.48% - ETA - 0 minutes                 [==================================================] - 99.24% - ETA - 0 minutes                 [==================================================] - 100% - ETA - 0 minutes                 
#>     sazo.effect D.level D.sazo  log.like      mae        rae      mse      mase
#> 216           1    0.96  0.995 -564.0946 13.89506 0.05106264 331.4577 0.4338099
#> 194           1    0.96  0.990 -564.2203 13.87442 0.05120442 326.9578 0.4331654
#> 238           1    0.96  1.000 -564.3520 13.93190 0.05095782 338.6151 0.4349599
#> 172           1    0.96  0.985 -564.6741 13.87102 0.05139446 324.7224 0.4330593
#> 196           1    0.98  0.990 -564.9152 14.15894 0.05179215 348.6237 0.4420483
#>     interval.score
#> 216       102.4692
#> 194       102.1615
#> 238       103.9692
#> 172       101.9692
#> 196       104.2154
#> Fitted DGLM with 1 outcomes.
#> 
#> distributions:
#>     outcomes: Poisson
#> 
#> Coeficients (smoothed) at time 144:
#>                 Estimate Std. Error   t value Pr(>|t|)
#> Var.Poly.Level   6.21228    0.01303 476.91183   <1e-12 *** 
#> Var.Poly.Slope   0.00901    0.00044  20.67807   <1e-12 *** 
#> Var.Sazo.Main.1 -0.15270    0.00835 -18.29165   <1e-12 *** 
#> Var.Sazo.Aux.1  -0.06600    0.00841  -7.84658   <1e-12 *** 
#> Var.Sazo.Main.2 -0.01370    0.00814  -1.68269    0.092   
#> Var.Sazo.Aux.2   0.08284    0.00829   9.98794   <1e-12 *** 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> ---
#> One-step-ahead prediction
#> Log-likelihood        : -564.09457
#> Interval Score        :  102.46923
#> Mean Abs. Scaled Error:    0.43381
#> Relative abs. Error   :    0.05106
#> Mean Abs. Error       :   13.89506
#> Mean Squared Error    :  331.45772
#> ---