Summary for a fitted kDGLM model

Prints a report for a fitted_dlm object.

Usage

# S3 method for fitted_dlm
summary(
  object,
  t = object$t,
  lag = -1,
  metric.lag = 1,
  metric.cutoff = floor(object$t/10),
  pred.cred = 0.95,
  ...
)

Arguments

object

A fitted_dlm object.

t

Integer: The time index for the latent states.

lag

Integer: The number of steps ahead used for the evaluating the latent states. Use lag<0 for the smoothed distribution, If lag==0 for the filtered distribution and lag=h for the h-step-ahead prediction.

metric.lag

Integer: The number of steps ahead used for the evaluating the predictions used when calculating metrics. Use metric.lag<0 for the smoothed distribution, If metric.lag==0 for the filtered distribution and metric.lag=h for the h-step-ahead prediction.

metric.cutoff

Integer: The cutoff time index for the metric calculation. Values before that time will be ignored.

pred.cred

numeric: The credibility interval to be used for the interval score.

...

Extra arguments passed to the coef method.#'

Value

No return value, called to print a summary of the fitted kDGLM model.

See also

Examples

data <- c(AirPassengers)

level <- polynomial_block(rate = 1, order = 2, D = 0.95)
season <- harmonic_block(rate = 1, order = 2, period = 12, D = 0.975)

outcome <- Poisson(lambda = "rate", data)

fitted.data <- fit_model(level, season,
  AirPassengers = outcome
)
summary(fitted.data)
#> Fitted DGLM with 1 outcomes.
#> 
#> distributions:
#>     AirPassengers: Poisson
#> 
#> ---
#> No static coeficients.
#> ---
#> See the coef.fitted_dlm for the coeficients with temporal dynamic.
#> 
#> One-step-ahead prediction
#> Log-likelihood        : -572.0896
#> Interval Score        : 106.83846
#> Mean Abs. Scaled Error:   0.40837
#> ---