update_Multinom_alt

Calculate the (approximated) posterior parameter for the linear predictors, assuming that the observed values came from a Multinomial model from which the logit probabilities have prior distribution in the log-Normal family.

Usage

update_Multinom_alt(conj.param, ft, Qt, y, parms = list())

Arguments

conj.param

list: A vector containing the concentration parameters of the Dirichlet. Not used in the alternative method.

ft

vector: A vector representing the means from the normal distribution.

Qt

matrix: A matrix representing the covariance matrix of the normal distribution.

y

vector: A vector containing the observations.

parms

list: A list of extra known parameters of the distribution. Not used in this kernel.

Value

The parameters of the posterior distribution.

Details

For evaluating the posterior parameters, we use a modified version of the method proposed in Alves et al. (2023) .

For computational efficiency, we also use a Laplace approximations to obtain the first and second moments of the posterior (see Tierney and Kadane 1986 and Tierney et al. 1989 ) .

For the details about the implementation see dos Santos Jr. et al. (?) .

For the detail about the modification of the method proposed in Alves et al. (2023) , see dos Santos Jr. et al. (?) .

References

Mariane Branco Alves, Helio S. Migon, Raíra Marotta, Silvaneo V. dos Santos Jr. au2 (2023). “k-parametric Dynamic Generalized Linear Models: a sequential approach via Information Geometry.” doi:10.48550/ARXIV.2201.05387 , https://arxiv.org/abs/2201.05387.

Silvaneo V. dos Santos Jr., Mariane Branco Alves, Helio S. Migon au2 (?). “kDGLM: an R package for Bayesian analysis of Dynamic Generialized Linear Models.” doi:? , ?, ?.

Silvaneo V. dos Santos Jr., Mariane Branco Alves, Helio S. Migon au2 (?). “AltMethod?” doi:? , ?, ?.

Luke Tierney, Joseph B. Kadane (1986). “Accurate Approximations for Posterior Moments and Marginal Densities.” Journal of the American Statistical Association, 81(393), 82-86. doi:10.1080/01621459.1986.10478240 , https://www.tandfonline.com/doi/pdf/10.1080/01621459.1986.10478240 , https://www.tandfonline.com/doi/abs/10.1080/01621459.1986.10478240 .

Luke Tierney, Robert E. Kass, Joseph B. Kadane (1989). “Fully Exponential Laplace Approximations to Expectations and Variances of Nonpositive Functions.” Journal of the American Statistical Association, 84(407), 710-716. doi:10.1080/01621459.1989.10478824 , https://doi.org/10.1080/01621459.1989.10478824 , https://doi.org/10.1080/01621459.1989.10478824.

See also