Search the best hyper parameters fo a kDGLM model

Search the best hyper parameters fo a kDGLM model

Usage

search_model(
  ...,
  search.grid,
  condition = "TRUE",
  metric = "log.like",
  smooth = TRUE,
  lag = 1,
  pred.cred = 0.95,
  metric.cutoff = NA,
  p.monit = NA,
  save.models = FALSE
)

Arguments

...

dlm_block or dlm_distr objects: The structural blocks of the model (dlm_block objects), alongside the model outcomes (dlm_distr object). At least one block must be undefined.

search.grid

list: A named list containing the possible values of each undefined hyper parameter.

condition

character: A character defining which combinations of undefined hyper parameter should be tested. See example for details.

metric

character: The name of the metric to use for model selection. One of log-likelihood for the one-step-ahead prediction ("log.like"), Mean Absolute Error ("mae"), Mean Absolute Scaled Error ("mase") (Hyndman and Koehler 2006) , Relative Absolute Error ("rae"), Mean Squared Error ("mse") or Interval Score ("interval.score") (Bracher et al. 2022) .

smooth

boolean: A flag indicating if the smoothed distribution for the latent variables should be calculated.

lag

integer: The number of steps ahead used for the prediction when calculating the metrics. If lag<0, predictions are made using the smoothed distribution of the latent variables.

pred.cred

numeric: A number between 0 and 1 (not included) indicating the credibility interval for predictions. If not within the valid range of values, 0.95 will be used.

metric.cutoff

integer: The number of observations to ignore when calculating the metrics. Default is 1/10 of the number of observations (rounded down).

p.monit

numeric (optional): The prior probability of changes in the latent space variables that are not part of its dynamic.

save.models

boolean: A flag indicating if all evaluated models should be saved. If FALSE, only the best model (according to the chosen metric) will be saved;

Value

A searched_dlm object containing the following values:

search.data data.frame: A table contain the metrics of each combination of hyper parameter tested and ordered from best combination to worst.

structure dlm_block object: The initial structure passed by the user, by with the undefined hyper parameters set to the best values found.

model fitted_dlm object: A model fitted with the best combination of hyper parameters.

Details

This is an auxiliary function to search for the best combination of hyper parameters among the possible variations specified by the user. This function simple evaluates all possible combinations and chooses the one that results in the best fitting.

References

Johannes Bracher, Evan L. Ray, Tilmann Gneiting, Nicholas G. Reich (2022). “Correction: Evaluating epidemic forecasts in an interval format.” PLOS Computational Biology, 18(10), 1-3. doi:10.1371/journal.pcbi.1010592 .

Rob J. Hyndman, Anne B. Koehler (2006). “Another look at measures of forecast accuracy.” International Journal of Forecasting, 22(4), 679--688. http://www.sciencedirect.com/science/article/B6V92-4K18VK5-1/1/227cef013b86100791c53712fae7677c.

See also

Examples


data <- c(AirPassengers)

level <- polynomial_block(rate = 1, order = 2, D = 0.95)
season <- harmonic_block(rate = "sazo.effect", order = 2, period = 12, D = "D.sazo")

outcome <- Poisson(lambda = "rate", data = data)

search_model(level, season,
  outcomes = outcome,
  search.grid = list(
    sazo.effect = c(0, 1),
    D.sazo = c(seq.int(0.95, 1, l = 11))
  ),
  condition = "sazo.effect==1 | D.sazo==1"
)
#> [====                                              ] - 8.33% - ETA - 0 minutes                 [========                                          ] - 16.67% - ETA - 0.01 minutes                 [============                                      ] - 25% - ETA - 0.01 minutes                 [=================                                 ] - 33.33% - ETA - 0.01 minutes                 [=====================                             ] - 41.67% - ETA - 0.01 minutes                 [=========================                         ] - 50% - ETA - 0.01 minutes                 [=============================                     ] - 58.33% - ETA - 0 minutes                 [=================================                 ] - 66.67% - ETA - 0 minutes                 [======================================            ] - 75% - ETA - 0 minutes                 [==========================================        ] - 83.33% - ETA - 0 minutes                 [==============================================    ] - 91.67% - ETA - 0 minutes                 [==================================================] - 100% - ETA - 0 minutes                 
#>    sazo.effect D.sazo  log.like      mae        rae      mse      mase
#> 20           1  0.995 -564.3869 13.84093 0.05111656 325.6502 0.4321200
#> 22           1  1.000 -564.5277 13.88405 0.05102099 332.5234 0.4334661
#> 18           1  0.990 -564.6147 13.83227 0.05130397 321.3583 0.4318495
#> 16           1  0.985 -565.1605 13.83729 0.05152992 319.2973 0.4320063
#> 14           1  0.980 -565.9742 13.85868 0.05180876 319.1286 0.4326739
#>    interval.score
#> 20       102.3692
#> 22       102.9538
#> 18       101.5154
#> 16       101.6769
#> 14       102.3462
#> Fitted DGLM with 1 outcomes.
#> 
#> distributions:
#>     outcomes: Poisson
#> 
#> Coeficients (smoothed) at time 144:
#>                 Estimate Std. Error   t value Pr(>|t|)
#> Var.Poly.Level   6.21075    0.01381 449.64055   <1e-12 *** 
#> Var.Poly.Slope   0.00897    0.00050  17.83528   <1e-12 *** 
#> Var.Sazo.Main.1 -0.15281    0.00840 -18.18911   <1e-12 *** 
#> Var.Sazo.Aux.1  -0.06640    0.00847  -7.83729   <1e-12 *** 
#> Var.Sazo.Main.2 -0.01352    0.00817  -1.65530    0.098   
#> Var.Sazo.Aux.2   0.08280    0.00832   9.95199   <1e-12 *** 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> ---
#> One-step-ahead prediction
#> Log-likelihood        : -564.38688
#> Interval Score        :  102.36923
#> Mean Abs. Scaled Error:    0.43212
#> Relative abs. Error   :    0.05112
#> Mean Abs. Error       :   13.84093
#> Mean Squared Error    :  325.65021
#> ---